Add 'current period' support to date parameters
