+++ frontend/src/metabase/meta/Parameter.js
@@ -60,11 +60,11 @@ type Deserializer = { testRegex: RegExp, deserialize: DeserializeFn}
 type DeserializeFn = (match: any[], fieldRef: LocalFieldReference | ForeignFieldReference) => FieldFilter;
 
 const timeParameterValueDeserializers: Deserializer[] = [
-    {testRegex: /^past([0-9]+)([a-z]+)s$/, deserialize: (matches, fieldRef) =>
-        ["time-interval", fieldRef, -parseInt(matches[0]), matches[1]]
+    {testRegex: /^past([0-9]+)([a-z]+)s(~)?$/, deserialize: (matches, fieldRef) =>
+        ["time-interval", fieldRef, -parseInt(matches[0]), matches[1], matches[2] ? { "include-current": true } : {}]
     },
-    {testRegex: /^next([0-9]+)([a-z]+)s$/, deserialize: (matches, fieldRef) =>
-        ["time-interval", fieldRef, parseInt(matches[0]), matches[1]]
+    {testRegex: /^next([0-9]+)([a-z]+)s(~)?$/, deserialize: (matches, fieldRef) =>
+        ["time-interval", fieldRef, parseInt(matches[0]), matches[1], matches[2] ? { "include-current": true } : {}]
     },
     {testRegex: /^this([a-z]+)$/, deserialize: (matches, fieldRef) =>
         ["time-interval", fieldRef, "current", matches[0]]
